requirements.txt
streamlit>=1.22
requests>=2.28
pandas>=2.0
scikit-learn>=1.2
numpy>=1.24
folium>=0.14
openai>=1.0.0
geopy>=2.4
pydeck>=0.8.0
python-dotenv>=1.0
streamlit-folium